Greenland Halibut/requirements.txt
python==3.10.0
tensorflow==2.11.0
pillow==9.3.0